Move websocket communication from config to event node
The communication with the websocket server on a Sony audio device has
been moved from the configuration node (device) to the event node which
is more natural as the event node is outputting the events. Also now the
connection to the websocket is only established if at least one event
node is used in a flow.